Move stores/file-system to file-system/controllers
