+++ src/components/diff-viewer/hooks/useDiffData.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from "react";
+import { useFileSystemStore } from "../../../file-system/controllers/store";
 import { useBufferStore } from "../../../stores/buffer-store";
-import { useFileSystemStore } from "../../../stores/file-system/store";
 import type { GitDiff } from "../../../utils/git";
 import { getFileDiff } from "../../../utils/git";
