+++ src/components/editor-footer.tsx
@@ -1,8 +1,8 @@
 import { AlertCircle, Download, Terminal as TerminalIcon } from "lucide-react";
 import { useUpdater } from "@/settings/hooks/use-updater";
+import { useFileSystemStore } from "../file-system/controllers/store";
 import { usePersistentSettingsStore } from "../settings/stores/persistent-settings-store";
 import { useBufferStore } from "../stores/buffer-store";
-import { useFileSystemStore } from "../stores/file-system/store";
 import { useGitStore } from "../stores/git-store";
 import { useUIState } from "../stores/ui-state-store";
 import { getFilenameFromPath, getLanguageFromFilename } from "../utils/file-utils";
