+++ src/components/editor/code-editor.tsx
@@ -1,5 +1,6 @@
 import type React from "react";
 import { useEffect, useMemo, useRef } from "react";
+import { useFileSystemStore } from "../../file-system/controllers/store";
 import { useEditorScroll } from "../../hooks/use-editor-scroll";
 import { useHover } from "../../hooks/use-hover";
 import { LspClient } from "../../lib/lsp/lsp-client";
@@ -10,7 +11,6 @@ import { useEditorCursorStore } from "../../stores/editor-cursor-store";
 import { useEditorInstanceStore } from "../../stores/editor-instance-store";
 import { useEditorSearchStore } from "../../stores/editor-search-store";
 import { useEditorSettingsStore } from "../../stores/editor-settings-store";
-import { useFileSystemStore } from "../../stores/file-system/store";
 import { useLspStore } from "../../stores/lsp-store";
 import FindBar from "../find-bar";
 import Breadcrumb from "./breadcrumb";
