+++ src/components/layout/main-sidebar.tsx
@@ -3,9 +3,9 @@ import type React from "react";
 import { useMemo } from "react";
 import { cn } from "@/utils/cn";
 import FileTree from "../../file-explorer/views/file-tree";
+import { useFileSystemStore } from "../../file-system/controllers/store";
 import { usePersistentSettingsStore } from "../../settings/stores/persistent-settings-store";
 import { useBufferStore } from "../../stores/buffer-store";
-import { useFileSystemStore } from "../../stores/file-system/store";
 import { useProjectStore } from "../../stores/project-store";
 import { useSidebarStore } from "../../stores/sidebar-store";
 import { useUIState } from "../../stores/ui-state-store";
